fix: rename files to remove the V1 tags
